requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
Jinja2==3.1.3
waitress==3.0.0
discord.py==2.3.2
python-dotenv==1.0.1
gunicorn==21.2.0
pymongo==4.7.2 